# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Levy Collection Practice Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matters:
    get:
      operationId: listMatters
      tags: [matters]
      summary: List matters
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: schemeId
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: assignedToId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of matters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatterSummary"
    post:
      operationId: createMatter
      tags: [matters]
      summary: Create matter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatterRequest"
      responses:
        "201":
          description: Matter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matter"
  /matters/{id}:
    get:
      operationId: getMatter
      tags: [matters]
      summary: Get matter details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matter with full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatterDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMatter
      tags: [matters]
      summary: Update matter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatterRequest"
      responses:
        "200":
          description: Updated matter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matter"
  /matters/{id}/stage:
    patch:
      operationId: updateMatterStage
      tags: [matters]
      summary: Advance matter to next stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStageRequest"
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matter"
        "400":
          description: Invalid stage transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /matters/{id}/history:
    get:
      operationId: getMatterHistory
      tags: [matters]
      summary: Get matter stage history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stage history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StageHistory"
  /matters/{id}/financials:
    get:
      operationId: getMatterFinancials
      tags: [matters]
      summary: Get matter financial summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"

  # DOCUMENTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Scheme:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        agentId:
          type: string
        agentName:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        levyAmount:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        matterCount:
          type: ["integer", "null"]
      required:
        - id
        - name
        - agentId
        - isActive
        - createdAt
...
    Debtor:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        idNumber:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        physicalAddress:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        matterCount:
          type: ["integer", "null"]
      required:
        - id
        - firstName
        - lastName
        - fullName
        - status
        - createdAt
...
    Matter:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        debtorId:
          type: string
        schemeId:
          type: string
        unit:
          type: string
        stage:
          type: string
        priority:
          type: string
        status:
          type: string
        capitalArrears:
          type: number
        interest:
          type: number
        legalCosts:
          type: number
        totalPaid:
          type: number
        interestFromDate:
          type: ["string", "null"]
        lodDate:
          type: ["string", "null"]
        s129Date:
          type: ["string", "null"]
        summonsDate:
          type: ["string", "null"]
        judgmentDate:
          type: ["string", "null"]
        writDate:
          type: ["string", "null"]
        saleDate:
          type: ["string", "null"]
        assignedToId:
          type: ["string", "null"]
        createdById:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - reference
        - debtorId
        - schemeId
        - unit
        - stage
        - priority
        - status
        - capitalArrears
        - interest
        - legalCosts
        - totalPaid
        - createdById
        - createdAt
        - updatedAt
    MatterSummary:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        debtorName:
          type: string
        schemeName:
          type: string
        agentName:
          type: ["string", "null"]
        unit:
          type: string
        stage:
          type: string
        priority:
          type: string
        status:
          type: string
        capitalArrears:
          type: number
        interest:
          type: number
        legalCosts:
          type: number
        totalPaid:
          type: number
        totalOutstanding:
          type: number
        assignedToName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - reference
        - debtorName
        - schemeName
        - unit
        - stage
        - priority
        - status
        - capitalArrears
        - interest
        - legalCosts
        - totalPaid
        - totalOutstanding
        - createdAt
        - updatedAt
    MatterDetail:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        debtor:
          $ref: "#/components/schemas/Debtor"
        scheme:
          $ref: "#/components/schemas/Scheme"
        unit:
          type: string
        stage:
          type: string
        priority:
          type: string
        status:
          type: string
        capitalArrears:
          type: number
        interest:
          type: number
        legalCosts:
          type: number
        totalPaid:
          type: number
        interestFromDate:
          type: ["string", "null"]
        lodDate:
          type: ["string", "null"]
        s129Date:
          type: ["string", "null"]
        summonsDate:
          type: ["string", "null"]
        judgmentDate:
          type: ["string", "null"]
        writDate:
          type: ["string", "null"]
        saleDate:
          type: ["string", "null"]
        assignedToId:
          type: ["string", "null"]
        assignedToName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        history:
          type: array
          items:
            $ref: "#/components/schemas/StageHistory"
      required:
        - id
        - reference
        - debtor
        - scheme
        - unit
        - stage
        - priority
        - status
        - capitalArrears
        - interest
        - legalCosts
        - totalPaid
        - createdAt
        - updatedAt
        - tasks
        - documents
        - payments
        - history
    CreateMatterRequest:
      type: object
      properties:
        debtorId:
          type: string
        schemeId:
          type: string
        unit:
          type: string
        priority:
          type: string
        capitalArrears:
          type: number
        legalCosts:
          type: number
        interestFromDate:
          type: ["string", "null"]
        assignedToId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - debtorId
        - schemeId
        - unit
        - priority
        - capitalArrears
    UpdateMatterRequest:
      type: object
      properties:
        unit:
          type: string
        priority:
          type: string
        status:
          type: string
        capitalArrears:
          type: number
        legalCosts:
          type: number
        interestFromDate:
          type: ["string", "null"]
        assignedToId:
          type: ["string", "null"]
        lodDate:
          type: ["string", "null"]
        s129Date:
          type: ["string", "null"]
        summonsDate:
          type: ["string", "null"]
        judgmentDate:
          type: ["string", "null"]
        writDate:
          type: ["string", "null"]
        saleDate:
          type: ["string", "null"]
    UpdateStageRequest:
      type: object
      properties:
        stage:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - stage
    StageHistory:
      type: object
      properties:
        id:
          type: string
        matterId:
          type: string
        fromStage:
          type: ["string", "null"]
        toStage:
          type: string
        changedByName:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - matterId
        - toStage
        - changedByName
        - createdAt
    FinancialSummary:
      type: object
      properties:
        matterId:
          type: string
        capitalArrears:
          type: number
        interestAccrued:
          type: number
        interestRate:
          type: number
        interestFromDate:
          type: ["string", "null"]
        interestDays:
          type: ["integer", "null"]
        interestPerDay:
          type: ["number", "null"]
        legalCosts:
          type: number
        totalPaid:
          type: number
        totalOutstanding:
          type: number
        runningBalance:
          type: number
      required:
        - matterId
        - capitalArrears
        - interestAccrued
        - interestRate
        - legalCosts
        - totalPaid
        - totalOutstanding
        - runningBalance
    Document:
      type: object
      properties:
        id:
          type: string
        matterId:
          type: string
        matterReference:
          type: ["string", "null"]
        docType:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        generatedByName:
          type: ["string", "null"]
        sentVia:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - matterId
        - docType
        - fileName
        - fileUrl
        - createdAt
...
    Payment:
      type: object
      properties:
        id:
          type: string
        matterId:
          type: string
        amount:
          type: number
        method:
          type: string
        receivedDate:
          type: string
        allocatedTo:
          type: ["string", "null"]
        receiptedByName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - matterId
        - amount
        - method
        - receivedDate
        - createdAt
...
    Task:
      type: object
      properties:
        id:
          type: string
        matterId:
          type: string
        matterReference:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        assigneeId:
          type: string
        assigneeName:
          type: ["string", "null"]
        isAutoGen:
          type: boolean
        completedAt:
          type: ["string", "null"]
        completionNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - matterId
        - title
        - status
        - priority
        - assigneeId
        - isAutoGen
        - createdAt
...